+++ src/main/java/org/gdg/zipte_gdg/domain/apt/Education.java
@@ -1,0 +1,21 @@
+package org.gdg.zipte_gdg.domain.apt;
+
+import jakarta.persistence.*;
+
+@Entity
+public class Education {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    private String description;
+
+    private String coordination;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Apt apt;
+
+}
