+++ src/main/java/org/gdg/zipte_gdg/security/jwt/util/CustomJWTException.java
@@ -1,4 +1,4 @@
-package org.gdg.zipte_gdg.security.util;
+package org.gdg.zipte_gdg.security.jwt.util;
 
 public class CustomJWTException extends Throwable {
     public CustomJWTException(String invalid) {
