+++ src/main/java/org/gdg/zipte_gdg/security/oauth/domain/PrincipalDetails.java
@@ -1,17 +1,18 @@
-package org.gdg.zipte_gdg.domain.oauth2;
+package org.gdg.zipte_gdg.security.oauth.domain;
 
 import lombok.RequiredArgsConstructor;
-import org.gdg.zipte_gdg.api.service.oauth.response.UserDTO;
+import org.gdg.zipte_gdg.security.oauth.service.response.UserDTO;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Map;
 
 @RequiredArgsConstructor
-public class PrincipalDetails implements OAuth2User, UserDetails {
+public class PrincipalDetails implements OAuth2User, UserDetails, Serializable {
     private final UserDTO userDTO;
 
     @Override
